Change thrown exception types to conform to Wagon
Changed the exception types thrown from some of the S3 interactions to better
conform to the expected Wagon behavior. This was done to work around the fact
that even though the Wagon API has a method to determine if a file exists, the
`deploy` operation does not call it before trying to get a file. In the case
that you are deploying a project to a repository for the first time, many
files won't already exist causing issues. The `SimpleStorageServiceWagon` now
throws `ResourceDoesNotExistException`s where it makes sense.

Currently the wagon probably throws the wrong type of exceptions in some cases
as its exception handling is very broad in some cases. This is a subject for
future improvement.