Artifactory build configuration
In order to enable publication of releases to Maven Central, the build has been updated to use the SpringSource Artifactory instance for publication.  This means that snapshots and releases will be published to repo.springsource.org and then can be promoted to Maven Central from there.

In addition to Artifactory changes, all dependency versions have been updated.