Add support for eu-west-1
Previously the wagon could not be used to access buckets with a
location constraint of eu-west-1 (issue #7).  However, it was able to
access buckets with a location constraint of EU.  According to the
Amazon documentation[1], EU is the 'official' S3 location constraint
for the Ireland datacenter.  The truth however, is that both eu-west-1
and EU are valid S3 location constraints for it. For example, when a
bucket is created with the AWS console[2], the bucket is assigned the
eu-west-1 location constraint.  However, when the bucket is created
using some other tools (for example, Transmit[3]), the bucket is
assigned the EU location constraint.  This is not documented anywhere,
and notably, the location constraint for the northern Virgina
datacenter must be US and cannot be us-east-1.

This change adds support for buckets with an eu-west-1 location
constraint.  It also adds support for buckets with a us-gov-west-1
location constraint (GovCloud), but does not test that this
functionality works.  At this time, it is not possible to create a
test bucket in GovCloud so there are no tests that verify the
functionality.

[1]: http://docs.aws.amazon.com/general/latest/gr/rande.html#s3_region
[2]: http://console.aws.amazon.com
[3]: http://panic.com/transmit/
[4]: http://aws.amazon.com/govcloud-us/