Use password instead of passphrase
Previously, the aws-maven wagon used the username and passphrase
facilities in the wagon API to contribute credentials.  At some point,
the implementation of the API changed such that it was no longer
possible to use that pairing together.  This change updates the wagon
to use username and password now, which can be used together.

[#14]